refactor: isStandalone 유틸 함수로 변경
